+++ src/index.js
@@ -1,9 +1,9 @@
 import { View, Text, TouchableOpacity } from 'react-native';
 
-import styled from './styled';
+import styled, { ThemeProvider, withTheme } from './styled';
 
 styled.View = styled(View, { name: 'styled(View)' });
 styled.Text = styled(Text);
 styled.Touchable = styled(TouchableOpacity, { name: 'styled(Touchable)' });
 
-export default styled;
+export { styled as default, ThemeProvider, withTheme };