+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+- Support theming.
+- Fixed a bug where `withComponent()` was showing the original component name.
+
 # 0.0.3
 
 - Initial release.